fix: [tag] The tag of Filemanager and Desktop is not synchronized
Both set and get tag infomation by DBus

Log: fix bug
Bug: https://pms.uniontech.com/bug-view-180653.html

+++ src/plugins/common/dfmplugin-tag/dfmplugin_tag_global.h
@@ -32,7 +32,7 @@
 
 DPTAG_BEGIN_NAMESPACE
 
-enum class TagActionType : uint8_t {
+enum class TagActionType : uint32_t {
     kGetAllTags = 1,
     kGetTagsThroughFile,
     kGetSameTagsOfDiffFiles,
